Remove `crossinline` declarations where possible (#45)
`crossinline` declarations are used when a callback passed to an `inline` method might escape the current execution context, which makes it impossible to use non-local returns. This is not the case for most methods here though, so the declaration can be removed and non-local returns can be used. This can simplify handling some cases:

```
val content = uct.fold(
  onLoading = { return },
  onError = { return },
  onContent = { it }
)
doSomething(content)
```